poc-backend/requirements.txt
# Web framework
fastapi==0.115.*
uvicorn[standard]==0.32.*

# Bot Framework
botbuilder-core==4.16.*
botbuilder-integration-aiohttp==4.16.*
aiohttp==3.10.*

# LangGraph + LangChain
langgraph==0.3.*
langchain-core==0.3.*
langsmith==0.2.*

# Azure services
azure-identity==1.19.*
azure-keyvault-secrets==4.9.*
azure-storage-blob==12.23.*
azure-ai-inference==1.0.*
pymongo==4.10.*
motor==3.6.*

# PDF generation
jinja2==3.1.*
weasyprint==63.*
markdown==3.7.*

# Config & utils
pydantic==2.10.*
pydantic-settings==2.7.*
python-dotenv==1.0.*

# Observability
opentelemetry-api==1.29.*
opentelemetry-sdk==1.29.*
azure-monitor-opentelemetry==1.6.*

# Dev/test
pytest==8.3.*
pytest-asyncio==0.25.*
httpx==0.28.*
